# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Movie Review API",
        "description": "API para plataforma de reseñas de películas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/directors": {
            "get": {
                "tags": [
                    "Directores"
                ],
                "summary": "Listar todos los directores",
                "operationId": "0232862a597e188b8d6b90bcff38654f",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Número de directores por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de directores obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "actual": {
                                                    "type": "integer"
                                                },
                                                "pages": {
                                                    "type": "integer"
                                                },
                                                "index": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Directores"
                ],
                "summary": "Crear nuevo director (Solo admin)",
                "operationId": "745ff928efedfec1f1958d43a96aeade",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "director_name"
                                ],
                                "properties": {
                                    "director_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Christopher Nolan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Director creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "director_id": {
                                            "type": "integer"
                                        },
                                        "director_name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado - Se requieren permisos de administrador"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/directors/{id}": {
            "get": {
                "tags": [
                    "Directores"
                ],
                "summary": "Mostrar un director específico",
                "operationId": "e1fe8469701666f99eaefb3cf4946316",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del director",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Director encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "director_id": {
                                            "type": "integer"
                                        },
                                        "director_name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Director no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Directores"
                ],
                "summary": "Actualizar un director",
                "operationId": "f077ea4bb08f19f10d07b83a10897496",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del director",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "director_name"
                                ],
                                "properties": {
                                    "director_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Christopher Nolan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Director actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "director_id": {
                                            "type": "integer"
                                        },
                                        "director_name": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Director no encontrado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Directores"
                ],
                "summary": "Eliminar un director (Solo admin)",
                "operationId": "c37df0e8f775c8408a0579a943904992",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del director",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Director eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Director eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado - Se requieren permisos de administrador"
                    },
                    "404": {
                        "description": "Director no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}